update
- move navbar.php to inc/navbar.inc.php
- profile.php: show username from session data in
- navbar.php: add doctor tabs
- navbar.php: add margins
- fix links paths in all pages

+++ controllers/Admin.controller.php
@@ -72,6 +72,10 @@ class AdminController {
     }
 
     public function deleteUser($userID) {
+        if ($userID == $_SESSION['user_id']) {
+            echo "Can't delete yourself!";
+            return;
+        }
         $user = new User($userID);
         $user->deleteUser();
         redirect("/");
